+++ docs/source/history.md
@@ -144,8 +144,8 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.2.6-alpha
 
-- Input:
-  - Improved: `StatesTable` and `Timeseries` now have option for treating NaN values (by zero weights)
+- Output:
+  - Improved: `FarmResultsEval` can now handle NaN data in Einstein sums
 - Turbine models:
   - New: `TableFactors`, multiplies variables with data from a two dimensional table
   - Generalized: `kTI`, now includes optional offset `kb`